[WIP] Address feedback on missing CSRF middleware (#52)
* Initial plan

* refactor: move fetchCsrfToken to lib/csrf.ts to break circular dependency


* fix: restore original log prefix for consistency


---------
